models/main_models/alfred/my_requirements.txt
ai2thor==2.5.1
aws-requests-auth==0.4.3
blinker==1.7.0
botocore==1.34.66
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.3
filelock==3.13.1
Flask==3.0.2
fsspec==2024.2.0
h5py==3.10.0
idna==3.6
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
lit==18.1.1
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
pandas==2.2.1
pillow==10.2.0
progressbar2==4.4.2
protobuf==3.20.0
python-dateutil==2.9.0.post0
python-utils==3.8.2
python-xlib==0.33
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
revtok==0.0.3
six==1.16.0
sympy==1.12
tensorboardX==1.8
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.66.2
triton==2.0.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
vocab==0.0.5
Werkzeug==3.0.1
zipp==3.18.0
